feat: Repo,Auditing

+++ src/main/java/com/dmadev/spring/dto/CompanyReadDto.java
@@ -1,4 +1,4 @@
 package com.dmadev.spring.dto;
 
-public record CompanyReadDto(Long id) {
+public record CompanyReadDto(Integer id) {
 }
